+++ packages/demand-bloc-payment/src/base.d.ts
@@ -1,0 +1,5 @@
+declare namespace braintree {
+  export interface Client {
+    teardown(): Promise<void> | void
+  }
+}