+++ packages/demand-bloc-payment/src/constants.ts
@@ -1,0 +1,9 @@
+export const creditCardType = 'CreditCard'
+
+export const defaultPaymentOptions = {
+  paymentCardsEnabled: false,
+}
+
+export const errors = {
+  verifyCardError: 'Verify card response does not meet requirements',
+}